backend/requirements.txt
connexion[flask] >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
gunicorn
flask
flask-cors
python-dotenv
pyjwt
mysql-connector-python

